+++ src/main/java/com/verdantartifice/primalmagick/client/fx/particles/InfernalFlameParticle.java
@@ -1,0 +1,68 @@
+package com.verdantartifice.primalmagick.client.fx.particles;
+
+import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.particle.Particle;
+import net.minecraft.client.particle.ParticleProvider;
+import net.minecraft.client.particle.ParticleRenderType;
+import net.minecraft.client.particle.RisingParticle;
+import net.minecraft.client.particle.SpriteSet;
+import net.minecraft.core.particles.SimpleParticleType;
+import net.minecraft.util.Mth;
+
+/**
+ * Particle type shown when calcinating items in an essence furnace.
+ * 
+ * @author Daedalus4096
+ */
+public class InfernalFlameParticle extends RisingParticle {
+    public InfernalFlameParticle(ClientLevel world, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed) {
+        super(world, x, y, z, xSpeed, ySpeed, zSpeed);
+    }
+
+    @Override
+    public ParticleRenderType getRenderType() {
+        return ParticleRenderType.PARTICLE_SHEET_OPAQUE;
+    }
+
+    @Override
+    public void move(double xd, double yd, double zd) {
+        this.setBoundingBox(this.getBoundingBox().move(xd, yd, zd));
+        this.setLocationFromBoundingbox();
+    }
+
+    @Override
+    public float getQuadSize(float partialTicks) {
+        float f = ((float)this.age + partialTicks) / (float)this.lifetime;
+        return this.quadSize * (1.0F - f * f * 0.5F);
+    }
+
+    @Override
+    protected int getLightColor(float partialTicks) {
+        float f = ((float)this.age + partialTicks) / (float)this.lifetime;
+        f = Mth.clamp(f, 0.0F, 1.0F);
+        int i = super.getLightColor(partialTicks);
+        int j = i & 255;
+        int k = i >> 16 & 255;
+        j += (int)(f * 15.0F * 16.0F);
+        if (j > 240) {
+            j = 240;
+        }
+
+        return j | k << 16;
+    }
+
+    public static class Factory implements ParticleProvider<SimpleParticleType> {
+        private final SpriteSet sprite;
+
+        public Factory(SpriteSet sprite) {
+            this.sprite = sprite;
+        }
+
+        @Override
+        public Particle createParticle(SimpleParticleType typeIn, ClientLevel worldIn, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed) {
+            InfernalFlameParticle ifp = new InfernalFlameParticle(worldIn, x, y, z, xSpeed, ySpeed, zSpeed);
+            ifp.pickSprite(this.sprite);
+            return ifp;
+        }
+    }
+}