+++ changelog.md
@@ -3,6 +3,7 @@ v2.1.1
 * Add a secret means to befriend witches, RIP Corspilla
 * Prevent hellish chain attacks from targeting team mates
 * Prevent smaller healing spell absorbs from applying when a larger absorb is already active
+* Update model for essence furnace
 
 v2.1.0-beta
 -----------